+++ cmake/opentelemetry-proto.cmake
@@ -2,17 +2,21 @@
 # SPDX-License-Identifier: Apache-2.0
 
 #
-# The dependency on opentelemetry-proto can be provided by order
-# of decreasing priority, options are:
+# The dependency on opentelemetry-proto can be provided by order of decreasing
+# priority, options are:
 #
-# 1 - Fetch from local source directory defined by the OTELCPP_PROTO_PATH variable
+# 1 - Fetch from local source directory defined by the OTELCPP_PROTO_PATH
+# variable
 #
-# 2 - Fetch from the opentelemetry-proto git submodule (opentelemetry-cpp/third_party/opentelemetry-proto)
+# 2 - Fetch from the opentelemetry-proto git submodule
+# (opentelemetry-cpp/third_party/opentelemetry-proto)
 #
-# 3 - Fetch from github using the git tag set in opentelemetry-cpp/third_party_release
+# 3 - Fetch from github using the git tag set in
+# opentelemetry-cpp/third_party_release
 #
 
-set(OPENTELEMETRY_PROTO_SUBMODULE "${opentelemetry-cpp_SOURCE_DIR}/third_party/opentelemetry-proto")
+set(OPENTELEMETRY_PROTO_SUBMODULE
+    "${opentelemetry-cpp_SOURCE_DIR}/third_party/opentelemetry-proto")
 
 if(OTELCPP_PROTO_PATH)
   if(NOT EXISTS
@@ -21,29 +25,32 @@ if(OTELCPP_PROTO_PATH)
       FATAL_ERROR
         "OTELCPP_PROTO_PATH does not point to a opentelemetry-proto repository")
   endif()
-  message(STATUS "fetching opentelemetry-proto from OTELCPP_PROTO_PATH=${OTELCPP_PROTO_PATH}")
-  FetchContent_Declare(
-      opentelemetry-proto
-      SOURCE_DIR ${OTELCPP_PROTO_PATH}
+  message(
+    STATUS
+      "fetching opentelemetry-proto from OTELCPP_PROTO_PATH=${OTELCPP_PROTO_PATH}"
   )
+  FetchContent_Declare(opentelemetry-proto SOURCE_DIR ${OTELCPP_PROTO_PATH})
   set(opentelemetry-proto_PROVIDER "fetch_source")
-  # If the opentelemetry-proto directory is a general directory then we don't have a good way to determine the version. Set it as unknown.
+  # If the opentelemetry-proto directory is a general directory then we don't
+  # have a good way to determine the version. Set it as unknown.
   set(opentelemetry-proto_VERSION "unknown")
 elseif(EXISTS ${OPENTELEMETRY_PROTO_SUBMODULE}/.git)
-   message(STATUS "fetching opentelemetry-proto from git submodule")
-   FetchContent_Declare(
-      opentelemetry-proto
-      SOURCE_DIR ${OPENTELEMETRY_PROTO_SUBMODULE}
-  )
+  message(STATUS "fetching opentelemetry-proto from git submodule")
+  FetchContent_Declare(opentelemetry-proto SOURCE_DIR
+                                           ${OPENTELEMETRY_PROTO_SUBMODULE})
   set(opentelemetry-proto_PROVIDER "fetch_source")
-  string(REGEX REPLACE "^v([0-9]+\\.[0-9]+\\.[0-9]+)$" "\\1" opentelemetry-proto_VERSION "${opentelemetry-proto_GIT_TAG}")
+  string(REGEX
+         REPLACE "^v([0-9]+\\.[0-9]+\\.[0-9]+)$" "\\1"
+                 opentelemetry-proto_VERSION "${opentelemetry-proto_GIT_TAG}")
 else()
   FetchContent_Declare(
-      opentelemetry-proto
-      GIT_REPOSITORY https://github.com/open-telemetry/opentelemetry-proto.git
-      GIT_TAG "${opentelemetry-proto_GIT_TAG}")
+    opentelemetry-proto
+    GIT_REPOSITORY https://github.com/open-telemetry/opentelemetry-proto.git
+    GIT_TAG "${opentelemetry-proto_GIT_TAG}")
   set(opentelemetry-proto_PROVIDER "fetch_repository")
-  string(REGEX REPLACE "^v([0-9]+\\.[0-9]+\\.[0-9]+)$" "\\1" opentelemetry-proto_VERSION "${opentelemetry-proto_GIT_TAG}")
+  string(REGEX
+         REPLACE "^v([0-9]+\\.[0-9]+\\.[0-9]+)$" "\\1"
+                 opentelemetry-proto_VERSION "${opentelemetry-proto_GIT_TAG}")
 endif()
 
 FetchContent_MakeAvailable(opentelemetry-proto)
@@ -174,8 +181,10 @@ foreach(IMPORT_DIR ${PROTOBUF_IMPORT_DIRS})
   list(APPEND PROTOBUF_INCLUDE_FLAGS "-I${IMPORT_DIR}")
 endforeach()
 
-set(PROTOBUF_COMMON_FLAGS "--proto_path=${PROTO_PATH}"
-                          "--cpp_out=${GENERATED_PROTOBUF_PATH}")
+set(PROTOBUF_COMMON_FLAGS
+    "--proto_path=${PROTO_PATH}"
+    "--cpp_out=dllexport_decl=OPENTELEMETRY_PROTO_API:${GENERATED_PROTOBUF_PATH}"
+)
 # --experimental_allow_proto3_optional is available from 3.13 and be stable and
 # enabled by default from 3.16
 if(Protobuf_VERSION AND Protobuf_VERSION VERSION_LESS "3.16")
@@ -276,13 +285,27 @@ add_custom_command(
   DEPENDS ${PROTOBUF_GENERATE_DEPENDS})
 
 unset(OTELCPP_PROTO_TARGET_OPTIONS)
-if(CMAKE_SYSTEM_NAME MATCHES "Windows|MinGW|WindowsStore")
-  list(APPEND OTELCPP_PROTO_TARGET_OPTIONS STATIC)
-elseif((NOT protobuf_lib_type STREQUAL "STATIC_LIBRARY")
-       AND (NOT DEFINED BUILD_SHARED_LIBS OR BUILD_SHARED_LIBS))
+if((NOT protobuf_lib_type STREQUAL "STATIC_LIBRARY")
+   AND (NOT DEFINED BUILD_SHARED_LIBS OR BUILD_SHARED_LIBS))
   list(APPEND OTELCPP_PROTO_TARGET_OPTIONS SHARED)
+  set(OTELCPP_PROTO_LIB_TYPE "SHARED_LIBRARY")
+
+  if(DEFINED BUILD_SHARED_LIBS AND NOT BUILD_SHARED_LIBS)
+    message(
+      WARNING
+        "Static protobuf libraries found, we force to building opentelemetry-proto as a static library to keep compatibility"
+    )
+  endif()
 else()
   list(APPEND OTELCPP_PROTO_TARGET_OPTIONS STATIC)
+  set(OTELCPP_PROTO_LIB_TYPE "STATIC_LIBRARY")
+
+  if(DEFINED BUILD_SHARED_LIBS AND BUILD_SHARED_LIBS)
+    message(
+      WARNING
+        "Shared protobuf libraries found, we force to building opentelemetry-proto as a shared library to keep compatibility"
+    )
+  endif()
 endif()
 
 set(OPENTELEMETRY_PROTO_TARGETS opentelemetry_proto)
@@ -299,13 +322,26 @@ add_library(
   ${METRICS_SERVICE_PB_CPP_FILE})
 set_target_version(opentelemetry_proto)
 
+if(OTELCPP_PROTO_LIB_TYPE STREQUAL "SHARED_LIBRARY")
+  project_build_tools_set_shared_library_declaration(OPENTELEMETRY_PROTO_API
+                                                     opentelemetry_proto)
+else()
+  project_build_tools_set_static_library_declaration(OPENTELEMETRY_PROTO_API
+                                                     opentelemetry_proto)
+endif()
+
 target_include_directories(
   opentelemetry_proto PUBLIC "$<BUILD_INTERFACE:${GENERATED_PROTOBUF_PATH}>"
                              "$<INSTALL_INTERFACE:include>")
 
 # Disable include-what-you-use and clang-tidy on generated code.
-set_target_properties(opentelemetry_proto PROPERTIES CXX_INCLUDE_WHAT_YOU_USE ""
-                                                     CXX_CLANG_TIDY "")
+set_target_properties(
+  opentelemetry_proto
+  PROPERTIES CXX_INCLUDE_WHAT_YOU_USE ""
+             CXX_CLANG_TIDY ""
+             C_VISIBILITY_PRESET "hidden"
+             CXX_VISIBILITY_PRESET "hidden"
+             VISIBILITY_INLINES_HIDDEN OFF)
 if(NOT Protobuf_INCLUDE_DIRS AND TARGET protobuf::libprotobuf)
   get_target_property(Protobuf_INCLUDE_DIRS protobuf::libprotobuf
                       INTERFACE_INCLUDE_DIRECTORIES)
@@ -323,10 +359,24 @@ if(WITH_OTLP_GRPC)
     ${LOGS_SERVICE_GRPC_PB_CPP_FILE} ${METRICS_SERVICE_GRPC_PB_CPP_FILE})
   set_target_version(opentelemetry_proto_grpc)
 
+  if(OTELCPP_PROTO_LIB_TYPE STREQUAL "SHARED_LIBRARY")
+    if(CMAKE_SYSTEM_NAME MATCHES "Windows|MinGW|WindowsStore")
+      # The codes generated by gRPC plugin do not support dll export/import
+      # declarations. To work around this, we enable WINDOWS_EXPORT_ALL_SYMBOLS
+      # property to export all symbols when building a shared library.
+      set_target_properties(opentelemetry_proto_grpc
+                            PROPERTIES WINDOWS_EXPORT_ALL_SYMBOLS ON)
+    endif()
+  endif()
+
   # Disable include-what-you-use and clang-tidy on generated code.
   set_target_properties(
-    opentelemetry_proto_grpc PROPERTIES CXX_INCLUDE_WHAT_YOU_USE ""
-                                        CXX_CLANG_TIDY "")
+    opentelemetry_proto_grpc
+    PROPERTIES CXX_INCLUDE_WHAT_YOU_USE ""
+               CXX_CLANG_TIDY ""
+               C_VISIBILITY_PRESET "default"
+               CXX_VISIBILITY_PRESET "default"
+               VISIBILITY_INLINES_HIDDEN OFF)
 
   list(APPEND OPENTELEMETRY_PROTO_TARGETS opentelemetry_proto_grpc)
   target_link_libraries(opentelemetry_proto_grpc PUBLIC opentelemetry_proto)
@@ -338,7 +388,8 @@ if(WITH_OTLP_GRPC)
   # opentelemetry_exporter_otlp_grpc_client as a static library.
   get_target_property(grpc_lib_type gRPC::grpc++ TYPE)
 
-  if(grpc_lib_type STREQUAL "SHARED_LIBRARY")
+  if(grpc_lib_type STREQUAL "SHARED_LIBRARY" OR OTELCPP_PROTO_LIB_TYPE STREQUAL
+                                                "SHARED_LIBRARY")
     target_link_libraries(opentelemetry_proto_grpc PUBLIC gRPC::grpc++)
   endif()
   set_target_properties(opentelemetry_proto_grpc PROPERTIES EXPORT_NAME